pyproject.toml
[tool.poetry]
name = "pytorch-caldera"
version = "0.0.1"
description = ""
homepage = "http://www.github.com/jvrana/caldera"
authors = ["Justin Vrana <justin.vrana@gmail.com>"]
classifiers = [
    'Intended Audience :: Developers',
    'Intended Audience :: Education',
    'Intended Audience :: Science/Research',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Topic :: Scientific/Engineering',
    'Topic :: Scientific/Engineering :: Mathematics',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Topic :: Software Development',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Python Modules'
]
keywords = ["pytorch", "machine learning", "network", "graph"]
packages = [
    {include="caldera"}
]

[tool.poetry.dependencies]
python = "^3.7"
networkx = "^2.4"
numpy = "^1.18.4"
scipy = "^1.5.2"
pytest-cov = {version = "^2.10.0", optional = true}
black = {version = "^19.10b0", optional = true}
pre-commit = {version = "^2.6.0", optional = true}
pylint = {version = "^2.5.3", optional = true}
sphinx_autodoc_typehints = {version = "^1.11.0", optional = true}
sphinx = {version = "^3.2.1", optional = true}
keats = {version = "^0.2.28", optional = true}
sphinx_bootstrap_theme = {version = "^0.7.1", optional = true}
nbformat = {version = "^5.0.7", optional = true}
nbconvert = {version = "^5.6.1", optional = true}
jupyter-sphinx = {version = "^0.3.1", optional = true}
pytest-sugar = {version = "^0.9.4", optional = true}
seaborn = {version = "^0.11.0", optional = true}
matplotlib = {version = "^3.3.1", optional = true}
tqdm = {version = "^4.48.2", optional = true}
pytest-randomly = "^3.4.1"
pytorch-lightning = {version = "^0.9.0", optional = true}
hydra-core = {version = "^1.0.0", optional = true}
attrs = "^20.2.0"
dill = "^0.3.2"
colorama = "^0.4.3"
rich = {version = "^7.0.0", optional = true}
h5py = {version = "^2.10.0", optional = true}
hydra-nevergrad-sweeper = {version = "^1.0.0-rc.5", optional = true}


[tool.poetry.extras]
docs = ["sphinx", "sphinx_bootstrap_theme", "sphinx_autodoc_typehints", "autodocsumm", "keats", "nbformat", "nbconvert", "jupyter-sphinx"]
lint = ["black", "pytest-cov", "pre-commit", "pylint"]
xtests = ["pytest-sugar"]
nbexamples = ["matplotlib", "seaborn", "tqdm", "pytorch-lightning", "hydra-core", "rich", 'h5py']

[tool.poetry.dev-dependencies]
pytest = "6"
flaky = "^3.7.0"
dictdiffer = "^0.8.1"
pytest-html = "^2.1.1"
pytest-benchmark = "^3.2.3"
pytest-xdist = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
